+++ src/Products/PloneMeeting/ftw_labels/overrides.py
@@ -2,33 +2,31 @@
 #
 # File: overrides.py
 #
-# Copyright (c) 2018 by Imio.be
-#
 # GNU General Public License (GPL)
 #
 
 from AccessControl import Unauthorized
 from ftw.labels.browser.labeling import Labeling
 from ftw.labels.browser.labelsjar import LabelsJar
+from ftw.labels.interfaces import ILabeling
+from ftw.labels.interfaces import ILabelSupport
 from ftw.labels.jar import LabelJar
-from ftw.labels.portlets.labeljar import Renderer as ftw_labels_renderer
+from ftw.labels.portlets.labeljar import Renderer as FTWLabelsRenderer
 from ftw.labels.viewlets.labeling import LabelingViewlet
 from plone import api
-from Products.CMFCore.permissions import ModifyPortalContent
-from Products.CMFCore.utils import _checkPermission
 from Products.PloneMeeting.config import PMMessageFactory as _
-from Products.PloneMeeting.utils import is_proposing_group_editor
 from Products.PloneMeeting.utils import notifyModifiedAndReindex
+from zope.annotation import IAnnotations
 
 
-class PMFTWLabelsRenderer(ftw_labels_renderer):
+class PMFTWLabelsRenderer(FTWLabelsRenderer):
     """ """
     @property
     def available(self):
         """ """
         available = super(PMFTWLabelsRenderer, self).available
         return available and \
-            self.context.getEnableLabels() and \
+            'labels' in self.context.getUsedItemAttributes() and \
             self.request.get('pageName', None) == 'data'
 
 
@@ -49,7 +47,8 @@ class PMLabelJar(LabelJar):
                   mapping={'item_url': brains[0].getURL()}),
                 type='error',
                 request=self.context.REQUEST)
-            return self.context.REQUEST.RESPONSE.redirect(self.context.REQUEST['HTTP_REFERER'])
+            return self.context.REQUEST.RESPONSE.redirect(
+                self.context.REQUEST['HTTP_REFERER'])
         notifyModifiedAndReindex(self.context)
         return super(PMLabelJar, self).remove(label_id)
 
@@ -70,6 +69,22 @@ class PMLabelsJar(LabelsJar):
 
 class PMLabeling(Labeling):
     """ """
+    def __init__(self, context, request):
+        super(PMLabeling, self).__init__(context, request)
+        self.tool = api.portal.get_tool('portal_plonemeeting')
+        self.cfg = self.tool.getMeetingConfig(self.context)
+
+    def available_labels(self, modes=('view', 'edit')):
+        # cache in request
+        cache_key = "ftw-labeling-cache-{0}-{1}".format(self.context.UID(), "-".join(modes))
+        cache = IAnnotations(self.request)
+        value = cache.get(cache_key, None)
+        if value is None:
+            labeling = ILabeling(self.context)
+            value = labeling.filter_manageable_labels(
+                labeling.available_labels(), modes=modes)
+            cache[cache_key] = value
+        return value
 
     def update(self):
         """ """
@@ -85,12 +100,7 @@ class PMLabeling(Labeling):
 
     @property
     def can_edit(self):
-        tool = api.portal.get_tool('portal_plonemeeting')
-        cfg = tool.getMeetingConfig(self.context)
-        return _checkPermission(ModifyPortalContent, self.context) or \
-            tool.isManager(cfg) or \
-            (cfg.getItemLabelsEditableByProposingGroupForever() and
-             is_proposing_group_editor(self.context.getProposingGroup(), cfg))
+        return bool(self.available_labels(modes=('edit', ))[1])
 
     @property
     def can_personal_edit(self):
@@ -103,15 +113,30 @@ class PMFTWLabelsLabelingViewlet(LabelingViewlet):
     def __init__(self, context, request, view, manager=None):
         super(PMFTWLabelsLabelingViewlet, self).__init__(context, request, view, manager=None)
         self.tool = api.portal.get_tool('portal_plonemeeting')
+        self.cfg = self.tool.getMeetingConfig(self.context)
 
     @property
     def available(self):
         """ """
-        available = super(PMFTWLabelsLabelingViewlet, self).available
-        if available:
-            cfg = self.tool.getMeetingConfig(self.context)
-            available = cfg.getEnableLabels()
-        return available
+        # for PloneMeeting
+        if 'labels' not in self.cfg.getUsedItemAttributes():
+            return False
+
+        # override to avoid several calls to self.available_labels
+        available_labels = self.available_labels
+        if not available_labels:
+            return False
+        if 'portal_factory' in self.context.absolute_url():
+            return False
+        if not ILabelSupport.providedBy(self.context):
+            return False
+        if not available_labels[0] and not available_labels[1]:
+            return False
+        return True
+
+    @property
+    def available_labels(self):
+        return self.context.restrictedTraverse('@@labeling').available_labels()
 
     @property
     def can_edit(self):