+++ src/Products/PloneMeeting/utils.py
@@ -37,6 +37,7 @@ from imio.helpers.content import get_user_fullname
 from imio.helpers.content import richtextval
 from imio.helpers.security import fplog
 from imio.helpers.workflow import get_final_states
+from imio.helpers.workflow import get_state_infos
 from imio.helpers.xhtml import addClassToContent
 from imio.helpers.xhtml import addClassToLastChildren
 from imio.helpers.xhtml import CLASS_TO_LAST_CHILDREN_NUMBER_OF_CHARS_DEFAULT
@@ -449,13 +450,13 @@ def _sendMail(obj, body, recipients, fromAddress, subject, format,
             for recipient in recipients:
                 obj.MailHost.send(
                     body, recipient, fromAddress, subject, charset='utf-8', msg_type=format)
-    except socket.error, sg:
+    except socket.error as sg:
         raise EmailError(SENDMAIL_ERROR % str(sg))
-    except UnicodeDecodeError, ue:
+    except UnicodeDecodeError as ue:
         raise EmailError(ENCODING_ERROR % str(ue))
-    except MailHostError, mhe:
+    except MailHostError as mhe:
         raise EmailError(MAILHOST_ERROR % str(mhe))
-    except Exception, e:
+    except Exception as e:
         raise EmailError(SENDMAIL_ERROR % str(e))
 
 
@@ -504,9 +505,9 @@ def sendMail(recipients, obj, event, attachments=None, mapping={}):
         for elt in mapping:
             if not isinstance(mapping[elt], unicode):
                 mapping[elt] = safe_unicode(mapping[elt])
-        translationMapping = mapping
+        translation_mapping = mapping
     else:
-        translationMapping = {}
+        translation_mapping = {}
 
     # get last WF action but specifically manage when an transition was
     # triggered automatilcally, the comments is in the previous transition
@@ -524,8 +525,8 @@ def sendMail(recipients, obj, event, attachments=None, mapping={}):
     else:
         # common case
         cfg_title = safe_unicode(cfg.Title())
-
-    translationMapping.update({
+    wf = api.portal.get_tool('portal_workflow').getWorkflowsFor(obj)[0]
+    translation_mapping.update({
         'portalUrl': portalUrl,
         'portalTitle': safe_unicode(portal.Title()),
         'objectTitle': safe_unicode(obj.Title()),
@@ -538,34 +539,35 @@ def sendMail(recipients, obj, event, attachments=None, mapping={}):
         'meetingConfigTitle': cfg_title,
         'transitionActor': wf_action and
         get_user_fullname(wf_action['actor'], with_user_id=True) or u'-',
-        'transitionTitle': wf_action and
-        translate(wf_action['action'], domain="plone", context=obj.REQUEST) or u'-',
+        'transitionTitle': translate(
+            safe_unicode(wf.transitions[wf_action['action']].title),
+            domain="plone",
+            context=obj.REQUEST) if (
+                wf_action and
+                wf_action['type'] == 'workflow' and
+                wf_action['action'] in wf.transitions) else u'-',
         'transitionComments': wf_action and safe_unicode(wf_action['comments']) or u'-',
     })
     if obj.getTagName() == 'Meeting':
-        translationMapping['meetingTitle'] = safe_unicode(obj.Title())
-        translationMapping['meetingLongTitle'] = tool.format_date(obj.date, prefixed=True)
-        translationMapping['meetingState'] = translate(obj.query_state(),
-                                                       domain='plone',
-                                                       context=obj.REQUEST)
+        translation_mapping['meetingTitle'] = safe_unicode(obj.Title())
+        translation_mapping['meetingLongTitle'] = tool.format_date(obj.date, prefixed=True)
+        translation_mapping['meetingState'] = get_state_infos(obj)['state_title']
     elif obj.getTagName() == 'MeetingItem':
-        translationMapping['itemTitle'] = safe_unicode(obj.Title())
-        translationMapping['itemState'] = translate(obj.query_state(),
-                                                    domain='plone',
-                                                    context=obj.REQUEST)
+        translation_mapping['itemTitle'] = safe_unicode(obj.Title())
+        translation_mapping['itemState'] = get_state_infos(obj)['state_title']
         meeting = obj.getMeeting()
         if meeting:
-            translationMapping['meetingUrl'] = get_public_url(meeting)
-            translationMapping['meetingTitle'] = safe_unicode(meeting.Title())
-            translationMapping['meetingLongTitle'] = tool.format_date(meeting.date, prefixed=True)
-            translationMapping['itemNumber'] = obj.getItemNumber(
+            translation_mapping['meetingUrl'] = get_public_url(meeting)
+            translation_mapping['meetingTitle'] = safe_unicode(meeting.Title())
+            translation_mapping['meetingLongTitle'] = tool.format_date(meeting.date, prefixed=True)
+            translation_mapping['itemNumber'] = obj.getItemNumber(
                 relativeTo='meeting')
 
     # some event end with "Owner", we use same event without the "Owner" suffix
     subjectLabel = u'%s_mail_subject' % event.replace("Owner", "")
     subject = translate(subjectLabel,
                         domain=d,
-                        mapping=translationMapping,
+                        mapping=translation_mapping,
                         context=obj.REQUEST)
     # special case for translations of event concerning state change
     # if we can not translate the specific translation msgid, we use a default msgid
@@ -579,14 +581,14 @@ def sendMail(recipients, obj, event, attachments=None, mapping={}):
             subjectLabel = u'item_state_changed_default_mail_subject'
         subject = translate(subjectLabel,
                             domain=d,
-                            mapping=translationMapping,
+                            mapping=translation_mapping,
                             context=obj.REQUEST)
     subject = safe_unicode(subject)
     # some event end with "Owner", we use same event without the "Owner" suffix
     bodyLabel = u'%s_mail_body' % event.replace("Owner", "")
     body = translate(bodyLabel,
                      domain=d,
-                     mapping=translationMapping,
+                     mapping=translation_mapping,
                      context=obj.REQUEST)
     # special case for translations of event concerning state change
     # if we can not translate the specific translation msgid, we use a default msgid
@@ -600,7 +602,7 @@ def sendMail(recipients, obj, event, attachments=None, mapping={}):
             bodyLabel = u'item_state_changed_default_mail_body'
         body = translate(bodyLabel,
                          domain=d,
-                         mapping=translationMapping,
+                         mapping=translation_mapping,
                          context=obj.REQUEST)
     body = safe_unicode(body)
 
@@ -625,16 +627,15 @@ def sendMail(recipients, obj, event, attachments=None, mapping={}):
         logger.info('Subject is [%s]' % subject)
         logger.info('Body is [%s]' % body)
         api.portal.show_message(extras, request=obj.REQUEST)
-        return obj, body, recipients, fromAddress, subject, attachments, translationMapping
     else:
         # Use 'plain' for mail format so the email client will turn links to clickable links
         mailFormat = 'text/plain'
         # Send the mail(s)
         try:
             _sendMail(obj, body, recipients, fromAddress, subject, mailFormat, attachments)
-        except EmailError, ee:
+        except EmailError as ee:
             logger.warn(str(ee))
-    return subject, body
+    return obj, body, recipients, fromAddress, subject, attachments, translation_mapping
 
 
 def sendMailIfRelevant(obj,
@@ -738,7 +739,7 @@ def sendMailIfRelevant(obj,
         recipient = getMailRecipient(user)
         # After all, we will add this guy to the list of recipients.
         recipients.append(recipient)
-    mail_subject = mail_body = None
+    subject = body = None
     if recipients:
         # wipeout recipients to avoid sendind same email to several users
         unique_emails = []
@@ -749,11 +750,12 @@ def sendMailIfRelevant(obj,
                 continue
             unique_emails.append(email)
             unique_email_recipients.append(recipient)
-        mail_subject, mail_body = sendMail(unique_email_recipients, obj, event, mapping=mapping)
+        obj, body, recipients, fromAddress, subject, attachments, translation_mapping = \
+            sendMail(unique_email_recipients, obj, event, mapping=mapping)
     debug = debug or obj.REQUEST.get('debug_sendMailIfRelevant', False)
     if debug:
-        obj.REQUEST.set('debug_sendMailIfRelevant_result', (recipients, mail_subject, mail_body))
-        return recipients, mail_subject, mail_body
+        obj.REQUEST.set('debug_sendMailIfRelevant_result', (recipients, subject, body))
+        return recipients, subject, body
     return True
 
 
@@ -826,6 +828,20 @@ def getDateFromDelta(aDate, delta):
     new_date = aDate + timedelta(int(days))
     new_date = new_date.replace(hour=int(hour), minute=int(minute))
     return new_date
+
+
+def is_operational_user(obj):
+    """Is current user an operationnal user in the application for the given p_obj."""
+    tool = api.portal.get_tool('portal_plonemeeting')
+    cfg = tool.getMeetingConfig(obj)
+    class_name = obj.__class__.__name__
+    return ((class_name == 'Meeting' and
+            _checkPermission(ModifyPortalContent, obj)) or
+         (not class_name == 'Meeting' and
+         (tool.isManager(cfg) or
+          bool(tool.userIsAmong(
+               suffixes=get_all_suffixes(omitted_suffixes=['observers']), cfg=cfg)))))
+
 
 
 def is_transition_before_date(obj, transition, date):
@@ -1315,7 +1331,7 @@ def applyOnTransitionFieldTransform(obj, transitionId):
       Apply onTransitionFieldTransforms defined in the corresponding obj MeetingConfig.
     '''
     idxs = []
-    extra_expr_ctx = _base_extra_expr_ctx(obj)
+    extra_expr_ctx = _base_extra_expr_ctx(obj, {'item': obj, })
     cfg = extra_expr_ctx['cfg']
     for transform in cfg.getOnTransitionFieldTransforms():
         tal_expr = transform['tal_expression'].strip()
@@ -1336,7 +1352,7 @@ def applyOnTransitionFieldTransform(obj, transitionId):
                     field = obj.getField(transform['field_name'].split('.')[1])
                     field.set(obj, res, mimetype='text/html')
                     idxs.append(field.accessor)
-            except Exception, e:
+            except Exception as e:
                 plone_utils = api.portal.get_tool('plone_utils')
                 plone_utils.addPortalMessage(
                     ON_TRANSITION_TRANSFORM_TAL_EXPR_ERROR % (
@@ -1356,7 +1372,7 @@ def meetingExecuteActionOnLinkedItems(meeting, transitionId, items=[]):
       check if we need to trigger an action on linked items
       defined in MeetingConfig.meetingExecuteActionOnLinkedItems.
     '''
-    extra_expr_ctx = _base_extra_expr_ctx(meeting)
+    extra_expr_ctx = _base_extra_expr_ctx(meeting, {'meeting': meeting, })
     cfg = extra_expr_ctx['cfg']
     wfTool = api.portal.get_tool('portal_workflow')
     wf_comment = _('wf_transition_triggered_by_application')
@@ -1378,7 +1394,7 @@ def meetingExecuteActionOnLinkedItems(meeting, transitionId, items=[]):
                     # do this as Manager to avoid permission problems, the configuration
                     # is supposed to be applied
                     with api.env.adopt_roles(['Manager']):
-                        extra_expr_ctx.update({'item': item})
+                        extra_expr_ctx.update({'item': item, })
                         _evaluateExpression(
                             item,
                             expression=action['tal_expression'].strip(),
@@ -1997,8 +2013,7 @@ def getAvailableMailingLists(obj, pod_template, include_recipients=False):
     if not mailing_lists:
         return res
     try:
-        extra_expr_ctx = _base_extra_expr_ctx(obj)
-        extra_expr_ctx.update({'obj': obj})
+        extra_expr_ctx = _base_extra_expr_ctx(obj, {'obj': obj, })
         for line in mailing_lists.split('\n'):
             name, expression, userIds = line.split(';')
             if not expression or _evaluateExpression(
@@ -2013,7 +2028,7 @@ def getAvailableMailingLists(obj, pod_template, include_recipients=False):
                     res.append((name, extract_recipients(obj, userIds)))
                 else:
                     res.append(name)
-    except Exception, exc:
+    except Exception as exc:
         msg = translate(
             'Mailing lists are not correctly defined, original error is \"${error}\"',
             domain='PloneMeeting',
@@ -2031,8 +2046,7 @@ def extract_recipients(obj, values):
     # compile userIds in case we have a TAL expression
     recipients = []
     userIdsOrEmailAddresses = []
-    extra_expr_ctx = _base_extra_expr_ctx(obj)
-    extra_expr_ctx.update({'obj': obj, })
+    extra_expr_ctx = _base_extra_expr_ctx(obj, {'obj': obj, })
     for value in values.strip().split(','):
         # value may be a TAL expression returning a list of userIds or email addresses
         # or a group (of users)
@@ -2358,7 +2372,7 @@ def org_id_to_uid(org_info, raise_on_error=True, ignore_underscore=False):
             org = getter(org_info.encode('utf-8'))
             if org:
                 return org.UID()
-    except Exception, exc:
+    except Exception as exc:
         if raise_on_error:
             raise(exc)
         else:
@@ -2444,7 +2458,7 @@ def get_next_meeting(meeting_date, cfg, date_gap=0):
     return res
 
 
-def _base_extra_expr_ctx(obj):
+def _base_extra_expr_ctx(obj, extra_ctx={}):
     """ """
     tool = api.portal.get_tool('portal_plonemeeting')
     cfg = tool.getMeetingConfig(obj)
@@ -2470,6 +2484,7 @@ def _base_extra_expr_ctx(obj):
             'imio_history_utils': SecureModuleImporter['imio.history.safe_utils'],
             'utils': SecureModuleImporter['Products.PloneMeeting.safe_utils'],
             'pm_utils': SecureModuleImporter['Products.PloneMeeting.safe_utils'], }
+    data.update(extra_ctx)
     return data